db用の関数を作成
